+++ python-libraries/data-platform-catalogue/data_platform_catalogue/client/datahub/search.py
@@ -2,7 +2,7 @@ import json
 import logging
 from datetime import datetime
 from importlib.resources import files
-from typing import Any, Literal, Sequence
+from typing import Any, Sequence
 
 from datahub.configuration.common import GraphError
 from datahub.ingestion.graph.client import DataHubGraph
@@ -161,6 +161,13 @@ class SearchClient:
         tags = self._parse_tags(entity)
         last_updated = self._parse_last_updated(entity)
         name = entity["name"]
+        relationships = entity.get("relationships", {})
+        total_data_products = relationships.get("total", 0)
+        data_products = relationships.get("relationships", [])
+        data_products = [
+            {"id": i["entity"]["urn"], "name": i["entity"]["properties"]["name"]}
+            for i in data_products
+        ]
 
         return SearchResult(
             id=entity["urn"],
@@ -171,6 +178,8 @@ class SearchClient:
             metadata={
                 "owner": owner_name,
                 "owner_email": owner_email,
+                "total_data_products": total_data_products,
+                "data_products": data_products,
             },
             tags=tags,
             last_updated=last_updated,
@@ -196,6 +205,7 @@ class SearchClient:
                 "owner": owner_name,
                 "owner_email": owner_email,
                 "domain": domain,
+                "number_of_assets": properties["numAssets"],
             },
             tags=tags,
             last_updated=last_updated,
@@ -203,10 +213,7 @@ class SearchClient:
 
     def _parse_facets(
         self, facets: list[dict[str, Any]]
-    ) -> dict[
-        Literal["domains", "tags", "customProperties", "glossaryTerms"],
-        list[FacetOption],
-    ]:
+    ) -> dict[str, list[FacetOption],]:
         """
         Parse the facets and aggregate information from the query results
         """