+++ python-libraries/data-platform-catalogue/tests/test_datahub_search.py
@@ -106,6 +106,7 @@ def test_one_search_result(mock_graph, searcher):
                     },
                     "owner": "",
                     "owner_email": "",
+                    "number_of_assets": 7,
                 },
                 tags=["custody"],
             )
@@ -176,7 +177,12 @@ def test_full_page(mock_graph, searcher):
                 result_type=ResultType.TABLE,
                 name="customers",
                 description="",
-                metadata={"owner": "", "owner_email": ""},
+                metadata={
+                    "owner": "",
+                    "owner_email": "",
+                    "data_products": [],
+                    "total_data_products": 0,
+                },
                 tags=[],
                 last_updated=datetime(2024, 1, 23, 6, 15, 2, 353000),
             ),
@@ -186,7 +192,12 @@ def test_full_page(mock_graph, searcher):
                 result_type=ResultType.TABLE,
                 name="customers2",
                 description="",
-                metadata={"owner": "", "owner_email": ""},
+                metadata={
+                    "owner": "",
+                    "owner_email": "",
+                    "data_products": [],
+                    "total_data_products": 0,
+                },
                 tags=[],
                 last_updated=None,
             ),
@@ -196,7 +207,12 @@ def test_full_page(mock_graph, searcher):
                 result_type=ResultType.TABLE,
                 name="customers3",
                 description="",
-                metadata={"owner": "", "owner_email": ""},
+                metadata={
+                    "owner": "",
+                    "owner_email": "",
+                    "data_products": [],
+                    "total_data_products": 0,
+                },
                 tags=[],
                 last_updated=None,
             ),
@@ -248,7 +264,12 @@ def test_query_match(mock_graph, searcher):
                 result_type=ResultType.TABLE,
                 name="customers",
                 description="",
-                metadata={"owner": "", "owner_email": ""},
+                metadata={
+                    "owner": "",
+                    "owner_email": "",
+                    "data_products": [],
+                    "total_data_products": 0,
+                },
                 tags=[],
             )
         ],
@@ -304,6 +325,8 @@ def test_result_with_owner(mock_graph, searcher):
                 metadata={
                     "owner": "Shannon Lovett",
                     "owner_email": "shannon@longtail.com",
+                    "data_products": [],
+                    "total_data_products": 0,
                 },
                 tags=[],
             )
@@ -414,4 +437,60 @@ def test_facets(searcher, mock_graph):
                 ),
             ],
         },
+    )
+
+
+def test_result_with_data_product(mock_graph, searcher):
+    datahub_response = {
+        "searchAcrossEntities": {
+            "start": 0,
+            "count": 1,
+            "total": 1,
+            "searchResults": [
+                {
+                    "entity": {
+                        "type": "DATASET",
+                        "urn": "urn:li:dataset:(urn:li:dataPlatform:bigquery,calm-pagoda-323403.jaffle_shop.customers,PROD)",  # noqa E501
+                        "name": "calm-pagoda-323403.jaffle_shop.customers",
+                        "relationships": {
+                            "total": 1,
+                            "relationships": [
+                                {
+                                    "entity": {
+                                        "urn": "urn:abc",
+                                        "properties": {"name": "abc"},
+                                    }
+                                }
+                            ],
+                        },
+                        "properties": {
+                            "name": "customers",
+                        },
+                    },
+                }
+            ],
+        }
+    }
+
+    mock_graph.execute_graphql = MagicMock(return_value=datahub_response)
+
+    response = searcher.search()
+    assert response == SearchResponse(
+        total_results=1,
+        page_results=[
+            SearchResult(
+                id="urn:li:dataset:(urn:li:dataPlatform:bigquery,calm-pagoda-323403.jaffle_shop.customers,PROD)",
+                matches={},
+                result_type=ResultType.TABLE,
+                name="customers",
+                description="",
+                metadata={
+                    "owner": "",
+                    "owner_email": "",
+                    "data_products": [{"id": "urn:abc", "name": "abc"}],
+                    "total_data_products": 1,
+                },
+                tags=[],
+            )
+        ],
     )