+++ python-libraries/data-platform-catalogue/CHANGELOG.md
@@ -7,6 +7,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] 2024-01-29
+
+### Added
+
+Enhanced the metadata returned with search results:
+
+- Added `number_of_assets` to data product metadata
+- Added `data_products` and `total_data_products` to dataset metadata
+
+### Changed
+
+- Replaced deprecated Datahub `filters` parameter with `orFilters`
+
 ## [0.7.0] 2024-01-25
 
 ### Added